+++ test/background_migrations/migration_job_test.rb
@@ -12,6 +12,7 @@ module BackgroundMigrations
       @connection.create_table(:users) do |t|
         t.boolean :admin
       end
+      User.reset_column_information
       EachBatchCalled.process_batch_called = 0
     end