+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 ## master (unreleased)
 
+- Add ability to reset counter caches using background migrations
+
+    ```ruby
+      class User < ApplicationRecord
+        has_many :projects
+      end
+
+      class Project < ApplicationRecord
+        belongs_to :user, counter_cache: true
+      end
+
+      class ResetUsersProjectsCount < ActiveRecord::Migration[7.0]
+        def up
+          reset_counters_in_background("User", :projects)
+        end
+      end
+    ```
+
 - Accept `0` as `batch_pause` value for background migrations
 - Ignore default scopes in `CopyColumn` and `BackfillColumn` background migrations
 - Raise an error for unsupported database versions